requirements.txt
streamlit
tensorflow==2.16.1
keras==3.3.3
segmentation-models
opencv-python
numpy
pillow